refactor: 홈 UI, 홈 뷰모델 수정 (#34)
