+++ presentation/home/src/main/java/com/yapp/breake/presentation/home/HomeScreen.kt
@@ -7,24 +7,21 @@ import androidx.compose.animation.togetherWith
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.yapp.breake.core.designsystem.theme.LinerGradient
+import com.yapp.breake.core.designsystem.theme.BackgroundGradient
 import com.yapp.breake.core.navigation.compositionlocal.LocalMainAction
 import com.yapp.breake.core.navigation.compositionlocal.LocalNavigatorAction
-import com.yapp.breake.presentation.HomeEvent
-import com.yapp.breake.presentation.HomeModalState
-import com.yapp.breake.presentation.HomeUiState
-import com.yapp.breake.presentation.HomeViewModel
 import com.yapp.breake.presentation.home.component.StopUsingDialog
 import com.yapp.breake.presentation.home.screen.BlockingScreen
 import com.yapp.breake.presentation.home.screen.ListScreen
@@ -38,7 +35,7 @@ internal fun HomeRoute(
 ) {
 	val homeUiState by viewModel.homeUiState.collectAsStateWithLifecycle()
 	val homeModalState by viewModel.homeModalState.collectAsStateWithLifecycle()
-	val mainNavigationAction = LocalNavigatorAction.current
+	val navAction = LocalNavigatorAction.current
 	val mainAction = LocalMainAction.current
 	val context = LocalContext.current
 
@@ -50,14 +47,14 @@ internal fun HomeRoute(
 		Box(
 			modifier = Modifier
 				.fillMaxWidth()
-				.fillMaxHeight(0.4f)
-				.background(brush = LinerGradient),
+				.height(360.dp)
+				.background(brush = BackgroundGradient),
 		)
 		HomeContent(
 			homeUiState = homeUiState,
 			viewModel = viewModel,
-			onShowAddScreen = mainNavigationAction::navigateToAddAppGroup,
-			onShowEditScreen = mainNavigationAction::navigateToEditAppGroup,
+			onShowAddScreen = viewModel::navigateToRegistry,
+			onShowEditScreen = viewModel::navigateToRegistry,
 		)
 	}
 
@@ -77,6 +74,9 @@ internal fun HomeRoute(
 							event.groupName,
 						),
 					)
+				}
+				is HomeEvent.NavigateToRegistry -> {
+					navAction.navigateToRegistry(groupId = event.groupId)
 				}
 			}
 		}
